Rename command get_rule and restore original get_rules name
