+++ pypanther/get_rule.py
