+++ tests/test_get.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 
 import pytest
 
-from pypanther.get import get_panther_rules, get_rules_from_module
+from pypanther.get import get_panther_rules, get_rules
 
 
 class TestGetPantherRules(TestCase):
@@ -38,18 +38,18 @@ class TestGetRulesFromModule(unittest.TestCase):
     def test_no_rules(self) -> None:
         from .fixtures.get_rules_test_data import no_rules
 
-        r = get_rules_from_module(module=no_rules)
+        r = get_rules(module=no_rules)
         assert len(r) == 0
 
     def test_rules(self) -> None:
         from .fixtures.get_rules_test_data import rules
 
-        r = get_rules_from_module(module=rules)
+        r = get_rules(module=rules)
         assert len(r) == 4
 
     def test_no_a_module(self) -> None:
         with pytest.raises(TypeError):
-            get_rules_from_module(module="str")
+            get_rules(module="str")
 
 
 if __name__ == "__main__":
